Refactor out zip function to common utils
